fix: fixing compilation errors
