--- src/routes.tsx
@@ -1,6 +1,0 @@
-const routes = [
-    {
-        name: 'home',
-        path: '/'
-    }
-]
